src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   watchfiles
    #   watchgod
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   django
billiard==4.2.0
    # via celery
celery==5.3.6
    # via
    #   -r requirements.in
    #   celery-redbeat
celery-redbeat==2.2.0
    # via -r requirements.in
certifi==2023.11.17
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   -r requirements.in
    #   cryptography
channels==4.0.0
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.2.0
    # via -r requirements.in
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==42.0.1
    # via
    #   django-anymail
    #   pyopenssl
    #   social-auth-core
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
django==4.2.9
    # via
    #   -r requirements.in
    #   channels
    #   django-anymail
    #   django-bootstrap-datepicker-plus
    #   django-debug-toolbar
    #   django-extensions
    #   social-auth-app-django
django-anymail==10.2
    # via -r requirements.in
django-bootstrap-datepicker-plus==5.0.5
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-webpack-loader==3.0.1
    # via -r requirements.in
docutils==0.20.1
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
freezegun==1.4.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   requests
kombu==5.3.5
    # via celery
markdown==3.5.2
    # via
    #   -r requirements.in
    #   markdown-urlize
markdown-urlize==0.2.0
    # via -r requirements.in
markupsafe==2.1.4
    # via werkzeug
msgpack==1.0.7
    # via channels-redis
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==23.2
    # via gunicorn
pillow==10.2.0
    # via -r requirements.in
prompt-toolkit==3.0.43
    # via click-repl
psycopg2==2.9.9
    # via -r requirements.in
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==2.5.3
    # via django-bootstrap-datepicker-plus
pydantic-core==2.14.6
    # via pydantic
pydot==2.0.0
    # via -r requirements.in
pyjwt==2.8.0
    # via social-auth-core
pyopenssl==24.0.0
    # via -r requirements.in
pyparsing==3.1.1
    # via
    #   -r requirements.in
    #   pydot
python-dateutil==2.8.2
    # via
    #   celery
    #   celery-redbeat
    #   freezegun
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.0.1
    # via uvicorn
python-jose==3.3.0
    # via -r requirements.in
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.1
    # via uvicorn
redis==5.0.1
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   channels-redis
requests==2.31.0
    # via
    #   -r requirements.in
    #   django-anymail
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rsa==4.9
    # via python-jose
sentry-sdk==1.39.2
    # via -r requirements.in
shortuuid==1.0.11
    # via -r requirements.in
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.0
    # via anyio
social-auth-app-django==5.4.0
    # via -r requirements.in
social-auth-core==4.5.1
    # via social-auth-app-django
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
tenacity==8.2.3
    # via celery-redbeat
typing-extensions==4.9.0
    # via
    #   django-bootstrap-datepicker-plus
    #   pydantic
    #   pydantic-core
tzdata==2023.4
    # via celery
urllib3==2.1.0
    # via
    #   django-anymail
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.27.0
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.21.0
    # via uvicorn
watchgod==0.8.2
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
websockets==12.0
    # via uvicorn
werkzeug==3.0.1
    # via -r requirements.in
whitenoise==6.6.0
    # via -r requirements.in
